backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
opencv-python==4.9.0.80
tensorflow
scikit-learn==1.4.0
nltk==3.8.1
numpy==1.26.3
joblib==1.3.2
python-multipart==0.0.6
pyjwt==2.8.0
bcrypt==4.1.2
email-validator
